requirements.txt
alembic==1.7.5
autopep8==1.6.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pycparser==2.21
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
Werkzeug==2.0.2
WTForms==3.0.0
